passing internal id to get event in eventEdit activity

+++ src/com/groupagendas/groupagenda/data/DataManagement.java
@@ -1181,7 +1181,7 @@ public class DataManagement {
 		return (items);
 	}
 
-	public Event getNativeCalendarEvent(int id) {
+	public Event getNativeCalendarEvent(long id) {
 		Event item = new Event();
 
 		Cursor cursor;
